error speech triggred only once
